backend/test/unit/modules/mapsmap/assets/delete_problem_asset.spec.ts
import { DeleteProblemAsset } from '../../../../../src/app/modules/mapsmap/assets/delete_problem_asset';

describe('DeleteProblemAsset', () => {
  let transactionAsset: DeleteProblemAsset;

	beforeEach(() => {
		transactionAsset = new DeleteProblemAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(3);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('deleteProblem');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('apply', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
